% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addup.r
\name{addnums}
\alias{addnums}
\alias{plot.funnypkg_addnums}
\title{The title of -addnums-
 
Here is a brief description}
\usage{
addnums(a, b)

\method{plot}{funnypkg_addnums}(x, y = NULL, ...)
}
\arguments{
\item{a}{Numeric scalar. A brief description.}

\item{b}{Numeric scalar. A brief description.}

\item{x}{An object of class \code{funnypkg_addnums}.}

\item{y}{Ignored.}

\item{...}{Further arguments passed to
\code{\link[graphics:plot.window]{plot.window}}.}
}
\value{
A list of class \code{funnypkg_addnums}:
\item{a}{Numeric scalar.}
\item{b}{Numeric scalar.}
\item{ab}{Numeric scalar. the sum of \code{a} and \code{b}}
}
\description{
The title of -addnums-
 
Here is a brief description
}
\details{
Computes the sum of \code{x} and \code{y}.
}
\examples{
addnums(1, 2)

}

