% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.r
\name{foo}
\alias{foo}
\alias{plot.funnypkg_foo}
\title{The title of -foo-}
\usage{
foo(a, b)

\method{plot}{funnypkg_foo}(x, y = NULL, ...)
}
\arguments{
\item{a}{Numeric scalar. A brief description.}

\item{b}{Numeric scalar. A brief description.}

\item{x}{An object of class \code{funnypkg_foo}.}

\item{y}{Ignored.}

\item{...}{Further arguments passed to
\code{\link[graphics:plot.window]{plot.window}}.}
}
\value{
A list of class \code{funnypkg_foo}:
\item{a}{Numeric scalar.}
\item{b}{Numeric scalar.}
\item{ab}{Numeric scalar. the sum of \code{a} and \code{b}}
}
\description{
The title of -foo-
}
\details{
Computes the sum of \code{x} and \code{y}.
}
\examples{
foo(1, 2)

}

