% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnypkg-pkg.r
\docType{package}
\name{funnypkg}
\alias{funnypkg}
\alias{funnypkg-package}
\title{funnypkg}
\description{
We add stuff up... You can access to the project
website at \url{http://github.com/USCBiostats/funnypkg}
}
\details{
A (not so) funny collection of functions
}
\author{
George G. Vega Yon
}

